+++ test/integration/Instana/SDK/IntegrationTest/WaiMiddleware.hs
@@ -12,6 +12,7 @@ import           Data.Aeson                             ((.=))
 import qualified Data.Aeson                             as Aeson
 import qualified Data.ByteString.Char8                  as BSC8
 import qualified Data.ByteString.Lazy.Char8             as LBSC8
+import           Data.List                              (isInfixOf)
 import qualified Data.List                              as List
 import           Data.Maybe                             (isNothing, listToMaybe)
 import           Instana.SDK.AgentStub.TraceRequest     (From (..), Span)
@@ -69,7 +70,7 @@ runTest pid urlPath headers extraAsserts = do
   response <-
     HttpHelper.doAppRequest Suite.testServerWithMiddleware urlPath "GET" headers
   let
-    result = LBSC8.unpack $ HTTP.responseBody response
+    responseBody = LBSC8.unpack $ HTTP.responseBody response
     from = Just $ From pid "agent-stub-id"
     responseHeaders = HTTP.responseHeaders response
     serverTimingTuple :: Maybe (Network.HTTP.Types.Header.HeaderName, BSC8.ByteString)
@@ -100,7 +101,7 @@ runTest pid urlPath headers extraAsserts = do
             Just entrySpan = maybeEntrySpan
             Just exitSpan = maybeExitSpan
           assertAllIO $
-            (commonAsserts entrySpan exitSpan result from serverTimingValue) ++
+            (commonAsserts entrySpan exitSpan responseBody from serverTimingValue) ++
             (extraAsserts entrySpan)
 
 
@@ -113,7 +114,7 @@ runSuppressedTest urlPath = do
       "GET"
       [("X-INSTANA-L", "0")]
   let
-    result = LBSC8.unpack $ HTTP.responseBody response
+    responseBody = LBSC8.unpack $ HTTP.responseBody response
   -- wait a second, then check that no spans have been recorded
   threadDelay $ 10 * 1000
   spansResults <-
@@ -126,7 +127,15 @@ runSuppressedTest urlPath = do
         then
           failIO "spans have been recorded although they should have not"   else
           assertAllIO
-            [ assertEqual "result" "{\"response\": \"ok\"}" result
+            [ assertBool
+                "downstream X-INSTANA-L"
+                (isInfixOf "\"X-INSTANA-L\":\"0\"" responseBody)
+            , assertBool
+                "no downstream X-INSTANA-T"
+                (not $ isInfixOf "X-INSTANA-T" responseBody)
+            , assertBool
+                "no downstream X-INSTANA-S"
+                (not $ isInfixOf "X-INSTANA-S" responseBody)
             ]
 
 
@@ -160,17 +169,30 @@ commonAsserts ::
   -> Maybe From
   -> Maybe String
   -> [Assertion]
-commonAsserts entrySpan exitSpan result from serverTimingValue =
-  [ assertEqual "result" "{\"response\": \"ok\"}" result
-  , assertEqual "trace ID is consistent"
-      (TraceRequest.t exitSpan)
-      (TraceRequest.t entrySpan)
-  , assertEqual "Server-Timing header with trace ID is present"
+commonAsserts entrySpan exitSpan responseBody from serverTimingValue =
+  [ assertEqual "Server-Timing header with trace ID is present"
       (Just $ "intid;desc=" ++ (TraceRequest.t entrySpan))
       (serverTimingValue)
   , assertEqual "exit parent ID"
       (Just $ TraceRequest.s entrySpan)
       (TraceRequest.p exitSpan)
+  , assertBool
+      ("wrong downstream X-INSTANA-T: " ++ responseBody ++
+          ", expected " ++ TraceRequest.t entrySpan)
+      (isInfixOf
+        ("\"X-INSTANA-T\":\"" ++ (TraceRequest.t entrySpan) ++ "\"")
+        responseBody
+      )
+  , assertBool
+      ("wrong downstream X-INSTANA-S: " ++ responseBody ++
+          ", expected " ++ TraceRequest.s exitSpan)
+      (isInfixOf
+        ("\"X-INSTANA-S\":\"" ++ (TraceRequest.s exitSpan) ++ "\"")
+        responseBody
+      )
+  , assertBool
+      "no downstream X-INSTANA-L"
+      (not $ isInfixOf "X-INSTANA-L" responseBody)
   , assertBool "entry timestamp" $ TraceRequest.ts entrySpan > 0
   , assertBool "entry duration" $ TraceRequest.d entrySpan > 0
   , assertEqual "entry kind" 1 (TraceRequest.k entrySpan)
@@ -185,7 +207,7 @@ commonAsserts entrySpan exitSpan result from serverTimingValue =
     ( Aeson.object
       [ "http" .= (Aeson.object
           [ "method" .= ("GET" :: String)
-          , "url"    .= ("http://127.0.0.1:1302/" :: String)
+          , "url"    .= ("http://127.0.0.1:1208/echo" :: String)
           , "params" .= ("some=query&parameters=2" :: String)
           , "status" .= (200 :: Int)
           ]
@@ -205,6 +227,7 @@ asserts entrySpan =
           , "host"   .= ("127.0.0.1:1207" :: String)
           , "url"    .= ("/api" :: String)
           , "params" .= ("some=query&parameters=1" :: String)
+          , "status" .= (200 :: Int)
           ]
         )
       ]